types.tsに型定義移動
